+++ app/src/main/java/com/keylesspalace/tusky/SplashActivity.kt
@@ -21,7 +21,7 @@ import androidx.appcompat.app.AppCompatActivity
 import com.keylesspalace.tusky.db.AccountManager
 import com.keylesspalace.tusky.di.Injectable
 
-import com.keylesspalace.tusky.util.NotificationHelper
+import com.keylesspalace.tusky.components.notifications.NotificationHelper
 import javax.inject.Inject
 
 class SplashActivity : AppCompatActivity(), Injectable {
