+++ app/src/main/java/com/keylesspalace/tusky/components/notifications/NotificationHelper.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License along with Tusky; if not,
  * see <http://www.gnu.org/licenses>. */
 
-package com.keylesspalace.tusky.util;
+package com.keylesspalace.tusky.components.notifications;
 
 import android.app.NotificationChannel;
 import android.app.NotificationChannelGroup;
